trustlayer-backend/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.31.0
google-generativeai==0.8.2
python-dotenv==1.0.1
